updated api get all contribution

+++ sunshine-backend/src/main/java/com/ute/sunshinebackend/dto/ContributionDto/TotalMoneyDto.java
@@ -2,7 +2,9 @@ package com.ute.sunshinebackend.dto.ContributionDto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TotalMoneyDto {
-    private long totalMoney = 0;
+    private BigDecimal totalMoney = BigDecimal.valueOf(0);
 }
